Renamed class to OutputVerifier
